+++ changelog.md
@@ -1,6 +1,12 @@
 
 
 # Changelog
+* 2020-07-24
+  * Fixed an error in the touch handlers 'tap start' function (nullpointerexception if empty area was tapped).
+  * Added CubicBezierPath.getClosestT(Vertex) function.
+  * Added BezierPath.getClosestT(Vertex) function.
+* 2020-07-22
+  * Fixed missing dragStart and dragEnd events in touch mode (after switching from Touchy to AlloyFinger).
 * 2020-07-21
   * Changed the safeMargeByKeys function: all keys, from source and destination, will be used now.
   * Changed the fetch.bool and fetch.number function: they will try a JSON.parse now on type mismatch (and use the fallback on failure of that).